Move coerce_to_datetime into corehq.util.dates
